changement int en byte

+++ AMOC_Module02_EssaisPortSerie/include/MorseBuzzer.h
@@ -4,10 +4,10 @@
 class MorseBuzzer : public Morse
 {
 public:
-MorseBuzzer(int p_duree, int p_BuzzerPin);
-void afficherPoint() const;
-void afficherTrait() const;
-void texteAEmettre(String p_texte) const;
+    MorseBuzzer(int p_duree, byte p_BuzzerPin);
+    void afficherPoint() const;
+    void afficherTrait() const;
+    void texteAEmettre(String p_texte) const;
 private:
-    int m_BuzzerPin;
+    byte m_BuzzerPin;
 };
