+++ AMOC_Module02_EssaisPortSerie/include/MorseDEL.h
@@ -4,10 +4,10 @@
 class MorseDEL : public Morse
 {
 public:
-MorseDEL(int p_duree, int p_DELPin);
+MorseDEL(int p_duree, byte p_DELPin);
 void afficherPoint() const;
 void afficherTrait() const;
 void texteAEmettre(String p_texte) const;
 private:
-    int m_DELPin;
+    byte m_DELPin;
 };
